Extract agent settings section
